# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: dnsaur API
  version: "1.0"
  description: >
    REST API for the dnsaur DNS/DHCP server. All paths are prefixed
    /api/v1. Authenticate with either the dnsaur_session cookie (set by
    POST /auth/login) or an Authorization: Bearer <token> API token
    (see /tokens). API tokens may have scope "read" (GET/HEAD only) or
    "write" (all methods); the session cookie always has full access.
    A read-scoped token used for a POST/PUT/PATCH/DELETE request is
    rejected with 403 {"error": "read-only token"} — see the 403
    response documented on each mutating operation below. Scope is
    enforced by method, with one exception on the read side: a
    read-scoped token gets TSIG secrets redacted (see the TSIGKey
    schema's secret_redacted field), because a GET that returns a
    signing credential in the clear is not something "read-only"
    should cover. The Authorization scheme name is matched
    case-insensitively per RFC 9110 §11.1.
    There is no separate "upstreams" resource: upstream DNS servers are
    configured via the "upstreams" key in /settings (a comma-separated
    list), alongside "upstream.strategy". Each entry is plain by default
    ("1.1.1.1:53") or, with a scheme, DNS-over-TLS
    ("tls://1.1.1.1:853#cloudflare-dns.com") or DNS-over-HTTPS
    ("https://1.1.1.1/dns-query#cloudflare-dns.com") — see the "upstreams"
    key under PUT /settings for the full grammar.
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dnsaur_session
      description: Session cookie set by POST /auth/login. HttpOnly, SameSite=Strict.
    bearerAuth:
      type: http
      scheme: bearer
      description: API token minted via POST /tokens, sent as "Authorization Bearer <token>".
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
paths:
...
  /queries:
    get:
      tags: [Queries]
      summary: Search the query log
      parameters:
        - name: from
          in: query
          schema: { type: integer, format: int64, description: unix ms, inclusive lower bound }
        - name: to
          in: query
          schema: { type: integer, format: int64, description: unix ms, inclusive upper bound }
        - name: client
          in: query
          schema: { type: string, description: exact client IP }
        - name: q
          in: query
          schema: { type: string, description: substring match on query name }
        - name: decision
          in: query
          schema: { type: string, enum: [blocked, authoritative, cached, stale, forwarded, error], description: how the query was answered; `authoritative` is an answer from a zone this server holds }
        - name: type
          in: query
          schema: { type: string, description: DNS query type, e.g. A, AAAA }
        - name: limit
          in: query
          schema: { type: integer, description: 0 defaults to 100, capped at 1000 }
        - name: offset
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: integer, format: int64 }
                    at: { type: integer, format: int64, description: unix ms }
                    instance_id: { type: string }
                    client_ip: { type: string }
                    client_id: { type: integer, format: int64 }
                    q_name: { type: string }
                    q_type: { type: string }
                    decision: { type: string }
                    rule_id: { type: integer, format: int64 }
                    list_id: { type: integer, format: int64 }
                    upstream: { type: string }
                    r_code: { type: string }
                    duration_ms: { type: integer, format: int64 }
                    matched:
                      type: string
                      description: >
                        The rule pattern or list entry that fired. Empty for
                        anything that was not blocked, and for rows logged
                        before the column existed.
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /queries/tail:
    get:
      tags: [Queries]
      summary: Live-tail the query log as server-sent events
      description: >
        Streams newly-logged query entries as they occur. Each SSE
        "data:" line is a JSON-encoded query log entry (same shape as
        GET /queries items). The connection stays open until the client
        disconnects.
      responses:
        "200":
          description: text/event-stream of query log entries
          content:
            text/event-stream:
              schema: { type: string }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: streaming unsupported
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: query log disabled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /stats/overview:
    get:
      tags: [Stats]
      summary: Aggregate query counts over a recent window
      parameters:
        - name: hours
          in: query
          schema:
            type: integer
            maximum: 8784
            description: >
              Lookback window. <=0 defaults to 24; anything above 8784
              (24 x 366) is clamped to it, since a larger window overflowed
              the duration arithmetic and asked for a span in the future.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer, format: int64 }
                  blocked: { type: integer, format: int64 }
                  cached: { type: integer, format: int64, description: cached + stale }
                  forwarded: { type: integer, format: int64 }
                  clients: { type: integer, format: int64, description: distinct client count }
                  dropped: { type: integer, format: int64, description: "query log entries discarded since start because the write buffer was full; non-zero means the counts above are undercounts" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /stats/timeline:
    get:
      tags: [Stats]
      summary: Query decision counts bucketed over time
      parameters:
        - name: hours
          in: query
          schema:
            type: integer
            maximum: 8784
            description: >
              Lookback window. <=0 defaults to 24; anything above 8784
              (24 x 366) is clamped to it, since a larger window overflowed
              the duration arithmetic and asked for a span in the future.
      responses:
        "200":
          description: OK, sorted ascending by bucket
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    bucket: { type: integer, format: int64, description: unix seconds bucket start }
                    decisions:
                      type: object
                      description: decision -> count
                      additionalProperties: { type: integer, format: int64 }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /stats/top:
    get:
      tags: [Stats]
      summary: Top-N values for a metric over a recent window
      parameters:
        - name: metric
          in: query
          required: true
          schema: { type: string, enum: [domain, blocked_domain, client] }
        - name: n
          in: query
          schema: { type: integer, description: "<=0 defaults to 10, capped at 100" }
        - name: hours
          in: query
          schema:
            type: integer
            maximum: 8784
            description: >
              Lookback window. <=0 defaults to 24; anything above 8784
              (24 x 366) is clamped to it, since a larger window overflowed
              the duration arithmetic and asked for a span in the future.
      responses:
        "200":
          description: OK, sorted descending by count
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    key: { type: string }
                    count: { type: integer, format: int64 }
        "400":
          description: metric must be domain, blocked_domain or client
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...